MetricsExporter/requirements.txt
psutil
prometheus_client
requests
kafka-python
docker